- added double-specialized qsort, currently just a copy from glibc - fixed some 64bit type mess in tests
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@3760 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ libgwyddion/gwytestser.c
@@ -184,7 +184,7 @@ gwy_test_ser_deserialize(const guchar *buffer,
     double theta, *radius = NULL;
     gchar *string = NULL;
     GObject *unit = NULL;
-    gsize rsize, ssize, usize;
+    guint32 rsize, ssize, usize;
     GwySerializeSpec spec[] = {
         { 'd', "theta", &theta, NULL, },
         { 'D', "radius", &radius, &rsize, },
